Fix inverted dependency
