+++ pymc_marketing/mmm/validating.py
@@ -18,8 +18,6 @@ from warnings import warn
 
 import pandas as pd
 
-from pymc_marketing.mmm.additive_effect import _validate_non_numeric_dtype
-
 __all__ = [
     "ValidateChannelColumns",
     "ValidateControlColumns",
@@ -28,6 +26,38 @@ __all__ = [
     "validation_method_X",
     "validation_method_y",
 ]
+
+
+def _validate_non_numeric_dtype(values, name: str) -> None:
+    """Validate that values are not numeric dtype (to prevent ambiguous date parsing).
+
+    Parameters
+    ----------
+    values : array-like
+        The values to validate
+    name : str
+        The name of the column/coordinate for error messages
+
+    Raises
+    ------
+    ValueError
+        If the values have numeric dtype (excluding empty arrays)
+    """
+    temp = pd.Series(values)
+
+    # Skip validation for empty arrays (they default to float64 but are not truly numeric)
+    if len(temp) == 0:
+        return
+
+    # Check if the values are numeric
+    if pd.api.types.is_numeric_dtype(temp.dtype):
+        raise ValueError(
+            f"'{name}' has numeric dtype ({temp.dtype}). "
+            "Date columns must have string or datetime dtype to avoid ambiguous date parsing. "
+            "For example, pd.to_datetime([0, 1, 2, 3]) would create dates starting from "
+            "January 1st 1970 with nanosecond intervals, which is likely not intended. "
+            "Please ensure your date column is properly formatted as strings or datetime objects."
+        )
 
 
 def validation_method_y(method: Callable) -> Callable:
